strands-agent/requirements.txt
strands-agents[openai]
strands-agents-tools[a2a_client]
dotenv
